remove otlphttpexporter's dependency on otlpreceiver as well
