+++ config/configgrpc/go.mod
@@ -119,3 +119,5 @@ replace go.opentelemetry.io/collector/config/confighttp => ../confighttp
 replace go.opentelemetry.io/collector/receiver/otlpreceiver => ../../receiver/otlpreceiver
 
 replace go.opentelemetry.io/collector/exporter/otlpexporter => ../../exporter/otlpexporter
+
+replace go.opentelemetry.io/collector/exporter/otlphttpexporter => ../../exporter/otlphttpexporter
