+++ config/confighttp/go.mod
@@ -98,3 +98,5 @@ replace go.opentelemetry.io/collector/exporter/otlpexporter => ../../exporter/ot
 replace go.opentelemetry.io/collector/config/configgrpc => ../configgrpc
 
 replace go.opentelemetry.io/collector/receiver/otlpreceiver => ../../receiver/otlpreceiver
+
+replace go.opentelemetry.io/collector/exporter/otlphttpexporter => ../../exporter/otlphttpexporter
