+++ connector/forwardconnector/go.mod
@@ -97,3 +97,5 @@ replace go.opentelemetry.io/collector/config/internal => ../../config/internal
 replace go.opentelemetry.io/collector/receiver/otlpreceiver => ../../receiver/otlpreceiver
 
 replace go.opentelemetry.io/collector/semconv => ../../semconv
+
+replace go.opentelemetry.io/collector/exporter/otlphttpexporter => ../../exporter/otlphttpexporter
