+++ consumer/go.mod
@@ -77,3 +77,5 @@ replace go.opentelemetry.io/collector/config/confignet => ../config/confignet
 replace go.opentelemetry.io/collector/receiver/otlpreceiver => ../receiver/otlpreceiver
 
 replace go.opentelemetry.io/collector/config/configtls => ../config/configtls
+
+replace go.opentelemetry.io/collector/exporter/otlphttpexporter => ../exporter/otlphttpexporter
